Q5   integrar a aplicação móvel com a api (#5)
* feat: dashboard consumption api

* feat: api consumption Produtores and Infinite scroll

* formating: refactoring of code

+++ queijo_finos_mobile/lib/models/DataInsight.dart
@@ -1,7 +1,7 @@
 class DataInsight {
-  double active;
-  double activeInContemplation;
-  double dropout;
+  int active;
+  int activeInContemplation;
+  int dropout;
 
   DataInsight({
     required this.active,
@@ -11,7 +11,7 @@ class DataInsight {
 
   // Método para calcular a porcentagem
   Map<String, double> calculatePercentages() {
-    double total = active + activeInContemplation + dropout;
+    int total = active + activeInContemplation + dropout;
     return {
       'active': (active / total) * 100,
       'activeInContemplation': (activeInContemplation / total) * 100,
